// src/shared/form/types.ts
export type InputTypes =
  | "number"
  | "button"
  | "time"
  | "image"
  | "text"
  | "checkbox"
  | "color"
  | "date"
  | "datetime-local"
  | "email"
  | "file"
  | "hidden"
  | "month"
  | "password"
  | "radio"
  | "range"
  | "select"
  | "fetchselect"
  | "textarea"
  | "searchselect"
  |"countryselect"
  ;


// export type InputTypes= "number" | "search" | "button" | "time" | "image" | 
// "text" | "checkbox" | "color" | "date" | "datetime-local" | "email" | "file" | 
// "hidden" | "month" | "password" | "radio" | "range" |"select"|"fetchselect" |"textarea"


export interface FormOptions {
  field_name: string;
  field_type:InputTypes;
  default_value?: string | number;
  placeholder?:string;
  required?:boolean;
  collection?:string;
  filter_key?:string;
  filter_list?:any[];
  filter_key_2?:string;
  editing?:boolean
  queryFn?:(props: QueryFnProps) => UseQueryResult<any, unknown>
  options?: { name: string; value: string }[];
  misc?: { coll_name: string };
}

export interface FilterParams{
data:any[]; keyword:string
}

export interface QueryFnProps{
keyword:string
key:string[]
}